+++ src/roboto/cli/command/args.py
@@ -31,7 +31,10 @@ def JsonFileOrStrType(arg):
 
 
 class KeyValuePairsAction(argparse.Action):
-    value_dict: dict[str, typing.Any] = {}
+    def __init__(self, option_strings, dest, parse_json=True, **kwargs):
+        self.parse_json = parse_json
+        self.value_dict = {}
+        super().__init__(option_strings, dest, **kwargs)
 
     def __call__(
         self,
@@ -58,16 +61,14 @@ class KeyValuePairsAction(argparse.Action):
                     )
                 key, value = parts
 
-                if key in self.value_dict:
-                    raise parser.error(
-                        f"Key '{key}' was defined multiple times for '{self.dest}'"
-                    )
-                # Attempt to parse the value to better handle numbers, booleans, etc
+                # Conditionally parse JSON based on the parameter
                 parsed_value = value
-                try:
-                    parsed_value = json.loads(value)
-                except json.decoder.JSONDecodeError:
-                    pass  # swallow
+                if self.parse_json:
+                    try:
+                        parsed_value = json.loads(value)
+                    except json.decoder.JSONDecodeError:
+                        pass
+
                 self.value_dict[key] = parsed_value
 
             setattr(namespace, self.dest, self.value_dict)